[feat] modified and simplified HomeViewController, got rid off swiftlint warning
