+++ App/Course App/Views/UIKit/HomeViewController.swift
@@ -13,7 +13,9 @@ import UIKit
 final class HomeViewController: UIViewController {
     let logger = Logger()
     
-    @IBOutlet var categoriesCollectionView: UICollectionView!
+    // swiftlint:disable implicitly_unwrapped_optional
+    var categoriesCollectionView: UICollectionView!
+    // swiftlint:enable implicitly_unwrapped_optional
     
     // MARK: DataSources
     typealias DataSource = UICollectionViewDiffableDataSource<SectionData, Joke>
@@ -25,5 +27,124 @@ final class HomeViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         setup()
+    }
+}
+
+// MARK: - UICollectionViewDataSource
+private extension HomeViewController {
+    func readData() {
+        dataProvider.$data.sink { [weak self] data in
+            self?.applySnapshot(data: data, animatingDifferences: true)
+        }
+        .store(in: &cancellables)
+    }
+    
+    func applySnapshot(data: [SectionData], animatingDifferences: Bool = true) {
+        guard dataSource.snapshot().numberOfSections == 0 else {
+            var snapshot = dataSource.snapshot()
+            
+            // swiftlint:disable force_unwrapping
+            snapshot.moveItem((data.first?.jokes.first)!, afterItem: (data.first?.jokes.last)!)
+            // swiftlint:enable force_unwrapping
+            
+            dataSource.apply(snapshot, animatingDifferences: animatingDifferences)
+            return
+        }
+        
+        var snapshot = Snapshot()
+        snapshot.appendSections(data)
+        
+        data.forEach { section in
+            snapshot.appendItems(section.jokes, toSection: section)
+        }
+        
+        dataSource.apply(snapshot, animatingDifferences: animatingDifferences)
+    }
+    
+    func makeDataSource() -> DataSource {
+        let dataSource = DataSource(collectionView: categoriesCollectionView) { collectionView, indexPath, _ in
+            let section = self.dataSource.snapshot().sectionIdentifiers[indexPath.section]
+            let imageCell: ImageCollectionViewCell = collectionView.dequeueReusableCell(for: indexPath)
+            imageCell.imageView.image = section.jokes[indexPath.item].image
+            return imageCell
+        }
+        
+        dataSource.supplementaryViewProvider = { collectionView, kind, indexPath in
+            guard kind == UICollectionView.elementKindSectionHeader else {
+                return nil
+            }
+            
+            let section = self.dataSource.snapshot().sectionIdentifiers[indexPath.section]
+            let labelCell: LabelCollectionViewCell = collectionView.dequeueSupplementaryView(ofKind: UICollectionView.elementKindSectionHeader, for: indexPath)
+            labelCell.nameLabel.text = section.title
+            return labelCell
+        }
+        
+        return dataSource
+    }
+}
+
+// MARK: - UICollectionViewDelegate
+extension HomeViewController: UICollectionViewDelegate {
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        logger.info("I have tapped \(indexPath)")
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
+        logger.info("will display \(indexPath)")
+    }
+}
+
+// MARK: - UI SETUP
+private extension HomeViewController {
+    func setup() {
+        setupCollectionView()
+        readData()
+    }
+    func setupCollectionView() {
+        categoriesCollectionView = UICollectionView(frame: view.bounds, collectionViewLayout: createCompositionalLayout())
+        categoriesCollectionView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        categoriesCollectionView.backgroundColor = .bg
+        categoriesCollectionView.delegate = self
+        categoriesCollectionView.register(ImageCollectionViewCell.self)
+        categoriesCollectionView.register(LabelCollectionViewCell.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader)
+        view.addSubview(categoriesCollectionView)
+    }
+}
+
+// MARK: - Layout
+private extension HomeViewController {
+    func createCompositionalLayout() -> UICollectionViewLayout {
+        let layout = UICollectionViewCompositionalLayout { sectionIndex, _ in
+            let section = self.dataProvider.data[sectionIndex]
+            
+            switch section.title {
+            default:
+                return self.createSection(with: section)
+            }
+        }
+        
+        let config = UICollectionViewCompositionalLayoutConfiguration()
+        config.interSectionSpacing = Constants.interSectionSpacing
+        layout.configuration = config
+        return layout
+    }
+    
+    func createSection(with section: SectionData) -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
+        
+        let layoutItem = NSCollectionLayoutItem(layoutSize: itemSize)
+        
+        let layoutGroupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(Constants.layoutWidth), heightDimension: .estimated(Constants.layoutGroupHeight))
+        let layoutGroup = NSCollectionLayoutGroup.horizontal(layoutSize: layoutGroupSize, subitems: [layoutItem])
+        
+        let layoutSection = NSCollectionLayoutSection(group: layoutGroup)
+        layoutSection.orthogonalScrollingBehavior = .groupPagingCentered
+        
+        let layoutSectionHeaderSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(Constants.layoutWidth), heightDimension: .estimated(Constants.layoutHeaderHeight))
+        let layoutSectionHeader = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: layoutSectionHeaderSize, elementKind: UICollectionView.elementKindSectionHeader, alignment: .top)
+        layoutSection.boundarySupplementaryItems = [layoutSectionHeader]
+        
+        return layoutSection
     }
 }
