+++ App/Course App/Views/UIKit/HomeViewController.xib
@@ -10,7 +10,6 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="HomeViewController" customModule="App_Course" customModuleProvider="target">
             <connections>
-                <outlet property="categoriesCollectionView" destination="zIc-gP-vgj" id="Ngt-fd-Qgg"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
         </placeholder>
@@ -18,26 +17,8 @@
         <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
             <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="zIc-gP-vgj">
-                    <rect key="frame" x="0.0" y="59" width="393" height="793"/>
-                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="yOn-0L-TzN">
-                        <size key="itemSize" width="128" height="128"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
-            </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="zIc-gP-vgj" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="GKg-y4-OOv"/>
-                <constraint firstAttribute="bottom" secondItem="zIc-gP-vgj" secondAttribute="bottom" id="Qqe-kh-gYS"/>
-                <constraint firstItem="zIc-gP-vgj" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="hUE-Xc-Aly"/>
-                <constraint firstItem="zIc-gP-vgj" firstAttribute="trailing" secondItem="fnl-2z-Ty3" secondAttribute="trailing" id="kOn-Ij-JnD"/>
-            </constraints>
             <point key="canvasLocation" x="121" y="21"/>
         </view>
     </objects>
